pyproject.toml
[tool.poetry]
name = "app-example"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
openai = "^1.47.0"
pymongo = "^4.9.1"
pymilvus = "^2.4.8"
psycopg2-binary = "^2.9.9"
azure-storage-blob = "^12.23.1"
gradio = "^5.1.0"
python-liquid = "^1.12.1"
langgraph = "^0.2.37"
langchain-openai = "^0.2.2"
deepeval = "^1.3.9"
markdown = "^3.7"
redis = "^5.1.1"
weasyprint = "^64.1"
html2docx = "^1.6.0"
pypandoc = "^1.15"
pdf2docx = "^0.5.8"

[tool.poetry.group.dev.dependencies]
watchdog = "4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"